+++ src/migration.rs
@@ -693,7 +693,7 @@ fn migrate_0_11_to_0_12<P: AsRef<Path>>(path: P, backup: P) -> Result<()> {
     store.backup(1)?;
 
     update_data_source_0_11_to_0_12(&store)?;
-    update_httpthreatfields_0_11_to_0_12(&store)?;
+    update_dgafields_httpthreatfields_0_11_to_0_12(&store)?;
 
     store.purge_old_backups(0)?;
     Ok(())
@@ -855,10 +855,10 @@ where
     Ok(())
 }
 
-// Update HttpThreatFields: migrate from 0.11 to 0.12
+// Update DgaFields, HttpThreatFields: migrate from 0.11 to 0.12
 #[allow(clippy::too_many_lines)]
-fn update_httpthreatfields_0_11_to_0_12(store: &crate::Store) -> Result<()> {
-    use crate::{event::HttpThreatFields, EventKind};
+fn update_dgafields_httpthreatfields_0_11_to_0_12(store: &crate::Store) -> Result<()> {
+    use crate::{event::DgaFields, event::HttpThreatFields, EventKind};
     use chrono::{DateTime, Utc};
     use num_traits::FromPrimitive;
 
@@ -933,6 +933,64 @@ fn update_httpthreatfields_0_11_to_0_12(store: &crate::Store) -> Result<()> {
         }
     }
 
+    #[derive(Deserialize, Serialize)]
+    struct OldDgaFields {
+        pub source: String,
+        pub src_addr: IpAddr,
+        pub src_port: u16,
+        pub dst_addr: IpAddr,
+        pub dst_port: u16,
+        pub proto: u8,
+        pub duration: i64,
+        pub method: String,
+        pub host: String,
+        pub uri: String,
+        pub referer: String,
+        pub version: String,
+        pub user_agent: String,
+        pub request_len: usize,
+        pub response_len: usize,
+        pub status_code: u16,
+        pub status_msg: String,
+        pub username: String,
+        pub password: String,
+        pub cookie: String,
+        pub content_encoding: String,
+        pub content_type: String,
+        pub cache_control: String,
+    }
+
+    impl From<OldDgaFields> for DgaFields {
+        fn from(input: OldDgaFields) -> Self {
+            Self {
+                source: input.source,
+                src_addr: input.src_addr,
+                src_port: input.src_port,
+                dst_addr: input.dst_addr,
+                dst_port: input.dst_port,
+                proto: input.proto,
+                duration: input.duration,
+                method: input.method,
+                host: input.host,
+                uri: input.uri,
+                referer: input.referer,
+                version: input.version,
+                user_agent: input.user_agent,
+                request_len: input.request_len,
+                response_len: input.response_len,
+                status_code: input.status_code,
+                status_msg: input.status_msg,
+                username: input.username,
+                password: input.password,
+                cookie: input.cookie,
+                content_encoding: input.content_encoding,
+                content_type: input.content_type,
+                cache_control: input.cache_control,
+                confidence: 0.0,
+            }
+        }
+    }
+
     let event_db = store.events();
     for item in event_db.raw_iter_forward() {
         let (k, v) = item.context("Failed to read events Database")?;
@@ -953,6 +1011,14 @@ fn update_httpthreatfields_0_11_to_0_12(store: &crate::Store) -> Result<()> {
                 };
                 let http_event: HttpThreatFields = fields.into();
                 let new = bincode::serialize(&http_event).unwrap_or_default();
+                event_db.update((&k, &v), (&k, &new))?;
+            }
+            EventKind::DomainGenerationAlgorithm => {
+                let Ok(fields) = bincode::deserialize::<OldDgaFields>(v.as_ref()) else {
+                    return Err(anyhow!("Failed to migrate events: Invalid Event value"));
+                };
+                let dga_event: DgaFields = fields.into();
+                let new = bincode::serialize(&dga_event).unwrap_or_default();
                 event_db.update((&k, &v), (&k, &new))?;
             }
             _ => continue,
@@ -1387,6 +1453,68 @@ mod tests {
 
         let event_db = settings.store.events();
         assert!(event_db.put(&message).is_ok());
+
+        #[derive(Deserialize, Serialize)]
+        struct OldDgaFields {
+            pub source: String,
+            pub src_addr: IpAddr,
+            pub src_port: u16,
+            pub dst_addr: IpAddr,
+            pub dst_port: u16,
+            pub proto: u8,
+            pub duration: i64,
+            pub method: String,
+            pub host: String,
+            pub uri: String,
+            pub referer: String,
+            pub version: String,
+            pub user_agent: String,
+            pub request_len: usize,
+            pub response_len: usize,
+            pub status_code: u16,
+            pub status_msg: String,
+            pub username: String,
+            pub password: String,
+            pub cookie: String,
+            pub content_encoding: String,
+            pub content_type: String,
+            pub cache_control: String,
+        }
+
+        let value = OldDgaFields {
+            source: "reveiw1".to_string(),
+            src_addr: "192.168.4.100".parse::<IpAddr>().unwrap(),
+            src_port: 40000,
+            dst_addr: "31.3.245.100".parse::<IpAddr>().unwrap(),
+            dst_port: 80,
+            proto: 10,
+            duration: Utc::now().timestamp_nanos(),
+            method: "GET".to_string(),
+            host: "example.com".to_string(),
+            uri: "/path/to/uri".to_string(),
+            referer: "-".to_string(),
+            version: "1.1".to_string(),
+            user_agent: "sample browser".to_string(),
+            request_len: 100,
+            response_len: 300,
+            status_code: 200,
+            status_msg: "Ok".to_string(),
+            username: "-".to_string(),
+            password: "-".to_string(),
+            cookie: "c.o.o.k.i.e".to_string(),
+            content_encoding: "text/html".to_string(),
+            content_type: "-".to_string(),
+            cache_control: "no-cache".to_string(),
+        };
+        let message = EventMessage {
+            time,
+            kind: EventKind::DomainGenerationAlgorithm,
+            fields: bincode::serialize(&value).unwrap_or_default(),
+        };
+
+        let event_db = settings.store.events();
+        assert!(event_db.put(&message).is_ok());
+
         let (db_dir, backup_dir) = settings.close();
 
         assert!(super::migrate_0_11_to_0_12(db_dir.path(), backup_dir.path()).is_ok());